+++ addon/query/query-object.js
@@ -19,11 +19,13 @@ export default class QueryObject {
    * @param expand {Object} Specifies the related entities to be included in line with retrieved entities.
    * @param select {Object} A specific set of properties for each requested entity.
    * @param primaryKeyName {Object} The name of primary key field for specified model name in modelName property.
-   * @param extend {Object} A additional computed set of properties for each requested entity. Based on properties in predicate and order, but not included in select and expand.
+   * @param extend {Object} An additional computed set of properties for each requested entity. Based on properties in predicate and order, but not included in select and expand.
+   * @param customQueryParams {Object} An additional query parameters.
+   * @param dataType {String} A data type for the request. Supported types: 'json' and 'blob'.
    * @class QueryObject
    * @constructor
    */
-  constructor(modelName, id, projectionName, predicate, order, top, skip, count, expand, select, primaryKeyName, extend) {
+  constructor(modelName, id, projectionName, predicate, order, top, skip, count, expand, select, primaryKeyName, extend, customQueryParams, dataType) {
     this.id = id;
     this.modelName = modelName;
     this.projectionName = projectionName;
@@ -36,5 +38,7 @@ export default class QueryObject {
     this.select = select;
     this.primaryKeyName = primaryKeyName;
     this.extend = extend;
+    this.customQueryParams = customQueryParams;
+    this.dataType = dataType;
   }
 }